Added confirmation to deletion process for items and groups
Also added a generic ConfirmPopup. Useful for future things. Learned that React attempts to evaluate out a function when you pass said function to a state variable's setter. To fix this, just return the function from a function to use the function as the value of the state variable. Easy, right? Check out App.js in this commit, line 16.